requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.1.1
gunicorn==23.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RPi.GPIO==0.7.1
serial==0.0.97
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.1.3
